pyproject.toml
[project]
name = "ragbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "llama-index-core (>=0.12.43,<0.13.0)",
    "llama-index-llms-openrouter (>=0.3.2,<0.4.0)",
    "torch (>=2.6.0,<3.0.0)",
    "llama-index-embeddings-huggingface (>=0.5.4,<0.6.0)",
    "streamlit (>=1.46.0,<2.0.0)",
    "pyvis (>=0.3.2,<0.4.0)",
]

[tool.poetry]
packages = [{include = "ragbot", from = "src"}]


[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu124"
priority = "supplemental"


[tool.poetry.dependencies]
torch = {source = "pytorch-gpu", version = "2.6.0"}
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
